@@ -1,4 +1,5 @@
-module.exports = {
+export default {
+  extensionsToTreatAsEsm: ['.ts', '.tsx'],
   transform: {
     '^.+\\.(t|j)sx?$': '@swc/jest',
   },
